# daily-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Daily API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api",
    "paths": {
...
        "/api/daily/statistics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "provides statistical data about a user's activity including likes, views, number of dailies written, current mood, streak, and a predefined topic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Get user statistics",
                "responses": {
                    "200": {
                        "description": "An object of statistics including likes, views, dailies written, mood, streak, and topic",
                        "schema": {
                            "$ref": "#/definitions/model.StatisticsDTO"
                        }
                    },
                    "400": {
                        "description": "bad request - error message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized - error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.StatisticsDTO": {
            "type": "object",
            "properties": {
                "dailiesWritten": {
                    "description": "Number of dailies written",
                    "type": "integer"
                },
                "date": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "likes": {
                    "description": "Number of likes received",
                    "type": "integer"
                },
                "mood": {
                    "description": "Current mood based on user's entries",
                    "type": "string"
                },
                "streak": {
                    "description": "Current streak of daily entries",
                    "type": "integer"
                },
                "topics": {
                    "description": "Currently focused topic",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "views": {
                    "description": "Number of views",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}